prep_cmr to image_transform
